chore: access config
Refactored the implementation of access configuration
